Reorganized directories, fixed memory bug, added scalarFill function
