--- projects/Expected_Mat_test_output.txt
@@ -1,49 +1,0 @@
-matrix m
-1, 2, 3, 4
-5, 6, 7, 8
-matrix n
-1, 2, 3
-matrix x
-1, 2, 3
-4, 5, 6
-7, 8, 9
-matrix y
-1, 2, 3, 4, 5, 6
-7, 8, 9, 10, 11, 12
-13, 14, 15, 16, 17, 18
-19, 20, 21, 22, 23, 24
-transpose of x
-1, 4, 7
-2, 5, 8
-3, 6, 9
-transpose of m
-1, 5
-2, 6
-3, 7
-4, 8
-n + x
-2, 4, 6
-5, 7, 9
-8, 10, 12
-transpose of n + x
-2, 3, 4
-6, 7, 8
-10, 11, 12
-transpose of x - transpose of n
-0, 3, 6
-0, 3, 6
-0, 3, 6
-matrix multiplication n ^ x
-30, 36, 42
-m.copy(newMat)
-1, 2, 3, 4
-5, 6, 7, 8
-m.roi(1,3)
-2, 3
-6, 7
--x.roi(1,-1,1)
--5, -6
--8, -9
-y.roi(2,5,1,3)
-9, 10, 11
-15, 16, 17
