+++ projects/Flood Fill/Expected_floodOutput.txt
@@ -1,0 +1,60 @@
+Map1 before flood fill:
+9, 9, 9, 9, 9, 9, 9, 9, 9
+9, 0, 0, 0, 9, 0, 0, 0, 9
+9, 0, 0, 0, 9, 0, 0, 0, 9
+9, 0, 0, 9, 0, 0, 0, 0, 9
+9, 9, 9, 0, 0, 0, 9, 9, 9
+9, 0, 0, 0, 0, 9, 0, 0, 9
+9, 0, 0, 0, 9, 0, 0, 0, 9
+9, 0, 0, 0, 9, 0, 0, 0, 9
+9, 9, 9, 9, 9, 9, 9, 9, 9
+Filling top-left with 5 (connectivity 8):
+9, 9, 9, 9, 9, 9, 9, 9, 9
+9, 5, 5, 5, 9, 5, 5, 5, 9
+9, 5, 5, 5, 9, 5, 5, 5, 9
+9, 5, 5, 9, 5, 5, 5, 5, 9
+9, 9, 9, 5, 5, 5, 9, 9, 9
+9, 5, 5, 5, 5, 9, 5, 5, 9
+9, 5, 5, 5, 9, 5, 5, 5, 9
+9, 5, 5, 5, 9, 5, 5, 5, 9
+9, 9, 9, 9, 9, 9, 9, 9, 9
+Filling center with 7 (connectivity 4):
+9, 9, 9, 9, 9, 9, 9, 9, 9
+9, 5, 5, 5, 9, 7, 7, 7, 9
+9, 5, 5, 5, 9, 7, 7, 7, 9
+9, 5, 5, 9, 7, 7, 7, 7, 9
+9, 9, 9, 7, 7, 7, 9, 9, 9
+9, 7, 7, 7, 7, 9, 5, 5, 9
+9, 7, 7, 7, 9, 5, 5, 5, 9
+9, 7, 7, 7, 9, 5, 5, 5, 9
+9, 9, 9, 9, 9, 9, 9, 9, 9
+Map2 before flood fill:
+9, 9, 9, 9, 9, 9, 9, 9, 9
+9, 0, 0, 0, 9, 1, 1, 1, 9
+9, 0, 0, 0, 2, 0, 0, 1, 9
+9, 0, 0, 2, 0, 0, 0, 1, 9
+9, 9, 2, 0, 0, 0, 2, 9, 9
+9, 0, 0, 0, 0, 2, 0, 0, 9
+9, 1, 0, 0, 2, 0, 0, 0, 9
+9, 1, 1, 1, 9, 0, 0, 0, 9
+9, 9, 9, 9, 9, 9, 9, 9, 9
+Map2 Custom filling center with 5 (threshold 1):
+9, 9, 9, 9, 9, 9, 9, 9, 9
+9, 0, 0, 0, 9, 5, 5, 5, 9
+9, 0, 0, 0, 2, 5, 5, 5, 9
+9, 0, 0, 2, 5, 5, 5, 5, 9
+9, 9, 2, 5, 5, 5, 2, 9, 9
+9, 5, 5, 5, 5, 2, 0, 0, 9
+9, 5, 5, 5, 2, 0, 0, 0, 9
+9, 5, 5, 5, 9, 0, 0, 0, 9
+9, 9, 9, 9, 9, 9, 9, 9, 9
+Map2 reset, then Custom filling center with 5 (threshold 3):
+9, 9, 9, 9, 9, 9, 9, 9, 9
+9, 5, 5, 5, 9, 5, 5, 5, 9
+9, 5, 5, 5, 5, 5, 5, 5, 9
+9, 5, 5, 5, 5, 5, 5, 5, 9
+9, 9, 5, 5, 5, 5, 5, 9, 9
+9, 5, 5, 5, 5, 5, 5, 5, 9
+9, 5, 5, 5, 5, 5, 5, 5, 9
+9, 5, 5, 5, 9, 5, 5, 5, 9
+9, 9, 9, 9, 9, 9, 9, 9, 9
